agent-core/pyproject.toml
[project]
name = "agent-core"
version = "0.1.0"
description = "Core system of agentic AI service"
authors = [{ name = "Pratik Ghodke", email = "pratik.ghodke@sarvaha.com" }]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "sentence-transformers (>=3.4.0,<4.0.0)",
    "groq (>=0.15.0,<0.16.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "psycopg (>=3.2.4,<4.0.0)",
    "agno (>=1.0.1,<2.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "playwright (>=1.50.0,<2.0.0)",
    "ollama (>=0.4.7,<0.5.0)",
    "sse-starlette (>=2.2.1,<3.0.0)",
    "instructor (>=1.7.7,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "crawl4ai (>=0.5.0.post8,<0.6.0)",
    "aio-pika (>=9.5.5,<10.0.0)",
    "google-genai (>=1.9.0,<2.0.0)",
    "pymilvus[model] (>=2.5.8,<3.0.0)",
    "keybert (>=0.9.0,<0.10.0)",
    "mcp (>=1.7.1,<2.0.0)",
    "mistralai (>=1.8.0,<2.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "boto3 (>=1.38.42,<2.0.0)",
    "langfuse (>=3.1.1,<4.0.0)",
    "opentelemetry-sdk (>=1.34.1,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.34.1,<2.0.0)",
    "openinference-instrumentation-agno (>=0.1.8,<0.2.0)",
    "openlit (>=1.34.23,<2.0.0)",
    "openinference-instrumentation-google-genai (>=0.1.4,<0.2.0)",
    "youtube-transcript-api (>=1.1.1,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "transformers (>=4.53.3,<5.0.0)",
    "onnx (>=1.18.0,<2.0.0)",
    "onnxruntime (>=1.22.1,<2.0.0)",
    "onnxruntime-tools (>=1.7.0,<2.0.0)",
    "scalar-fastapi (>=1.2.2,<2.0.0)"
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.3"
pre-commit = "^4.1.0"
streamlit = "^1.42.2"

[tool.poetry]
package-mode = true
packages = [{ include = "app" }]

[tool.poetry.scripts]
"start:dev" = "run:run_dev"
"start:prod" = "run:run"
"run:celery" = "scripts:run_celery"
"run:celery:debug" = "scripts:run_celery_debug"
"run:pre-commit" = "scripts:run_pre_commit"
"run:linter" = "scripts:ruff_check"
"run:linter:fix" = "scripts:ruff_check_fix"
"run:quantize-reranker" = "scripts:quantize_reranker_model"


[tool.ruff]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    "PL",
    "E",
    "F",
    "W",
    "I",
    "N",
    "D",
    "SIM",
    "B",
    "A",
    "C4",
    "PIE",
    "RUF",
    "PERF",
    "ISC",
    "PIE",
    "RET",
    "UP",
    "PYI",
    "TRY",
]
ignore = ["TRY003"]

[tool.ruff.per-file-ignores]
"app/common/prompts.py" = ["E501"]
"app/common/constants.py" = ["E501"]
"app/middlewares/__init__.py" = ["TRY301"]


# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.mypy]
mypy_path = ["."]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
